third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/MutationObserver/removed-out-of-order-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Test MutationEvents interfering with MutationObservers: removing nodes 'out of order'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutations.length is 2
PASS mutations[0].addedNodes.length is 2
PASS mutations[0].removedNodes.length is 0
PASS mutations[0].addedNodes[0].tagName is 'B'
PASS mutations[1].addedNodes.length is 0
PASS mutations[1].removedNodes.length is 1
PASS mutations[1].removedNodes[0].tagName is 'B'
PASS mutations[1].addedNodes[0] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
